Testcase for StopWords reorg
